Overland/soil and overland/groundwater coupling flux calculation surface structure and much cleaning Overland flow CVFEM alternative saturation calculation for richards flow additional case for VanGenuchten soil water characteristic curves. Separate parameters for saturation-pressure, relative permeability-pressure relationship
git-svn-id: https://svn.ufz.de/svn/ogs/trunk/sources@1029 3895a583-e809-0410-9bb6-aa9463026377

+++ FEM/rf_apl.cpp
@@ -165,7 +165,7 @@ int RFPre_FEM(char* dateiname)
   //......................................................................
   DOMRead(dateiname);
   #ifdef RFW_FRACTURE
-  for(int i=0; i<mmp_vector.size(); ++i)
+  for(int i=0; i<(int)mmp_vector.size(); ++i)
     {
       if( mmp_vector[i]->frac_num >0)
       {
@@ -396,7 +396,7 @@ int ExecuteRFTimeLoop(void)
     /* Ergebnisausgabe */
     OUTData(m_tim->time_current,aktueller_zeitschritt);
     #ifdef RFW_FRACTURE 
-      for(int i=0; i<mmp_vector.size(); ++i)
+      for(int i=0; i<(int)mmp_vector.size(); ++i)
         {
           if( mmp_vector[i]->frac_num >0)
           {
